+++ src/parser.h
@@ -14,10 +14,10 @@ typedef struct {
 
 Token   *parser_peek_token(Parser *parser);
 Token   *parser_next_token(Parser *parser);
-void     print_id_mask(u32 id_mask, Str lexeme, FILE *stream);
+void     print_id_mask(u64 id_mask, Str lexeme, FILE *stream);
 void     expect_token(Token *token, u64 id_mask);
 Token   *parser_expect_token(Parser *parser, u64 id_mask);
-ArgKind  token_id_to_arg_kind(u32 id);
+ArgKind  token_id_to_arg_kind(u64 id);
 ValueKind str_to_value_kind(Str str);
 Value str_to_number_value(Str str);
 Arg str_to_arg(Str text, ArgKind kind);
